Implement word wrapping.
Holy crap this was hell. Had to do two fixes to markdown.php itself:
1) Allow markdown to wordwrap (insert <br /> where it finds \n) when
there aren't two spaces at the end of the preceeding line. Just
wordwarps at every "\n".

2) Add wordwrap call to automatically wrap text. I refuse to manually
wrap text as I'm writing because this is an insane and stupid amount of
management as you're typing, exepcially because META-Q fails in Sublime
(e.g. adds bullets to everything in your paragph if you're in a bullet).